* Merge everything into one .py file * Add fieldnames method, remove slotted method * Fix everything to note it's all in one file
