+++ performance/dataclass_serializers_compared_noorjson.py
@@ -10,7 +10,7 @@ import dataclasses
 
 import json
 
-from json_defaults.dataclasses import dataclass_default
+from json_defaults import dataclass_default
 
 
 ITERATIONS = 100
